// src/permissions/domain/permission.enum.ts
export enum PermissionModeEnum {
  CREATE = 'create',
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete',
  MANAGE = 'manage',
}

export enum PermissionEntityEnum {
  RoleEntity = 'RoleEntity',
  UserEntity = 'UserEntity',
  ClientEntity = 'ClientEntity',
  PermissionEntity = 'PermissionEntity',
  VisitEntity = 'VisitEntity',
}